Fix line wrapping for range function
This requires passing an extra indent len to the one_line_or_many func. In
most cases this is 0, but for range functions we have two separate pieces that
may need wrapping. The first is the func call and its args. The second is the
column definition list. The latter is already indented by however long the
last line of the first part is, but we don't want to push this onto the indent
stack, as we don't actually want to indent the second part by that
amount.