Add a macro for pushing a context and creating a guard to pop the context stack
The format_range_var_table_sample method was pushing a context without
popping. Having a macro that ensures both are always done simplifies this. I
originally wanted this to be a method that returned a guard, but I'm not sure
how to specify the ScopeGuard<...> type parameters in the return type. See
https://github.com/bluss/scopeguard/issues/13 for discussion of this.